More armor work (maybe less broken, maybe more)

+++ common/net/silentchaos512/gems/api/tool/part/ToolPart.java
@@ -179,7 +179,7 @@ public abstract class ToolPart {
   public abstract float getChargeSpeed();
 
   // Only needed for armor.
-  public abstract int getProtection();
+  public abstract float getProtection();
 
   /**
    * Determines if the part is valid for tools of the given tier. Generally, this is true if the part and tool tier is
