+++ common/net/silentchaos512/gems/api/tool/part/ToolPartGrip.java
@@ -66,7 +66,7 @@ public class ToolPartGrip extends ToolPart {
   }
 
   @Override
-  public final int getProtection() {
+  public final float getProtection() {
 
     return 0;
   }
